requirements.txt
opencv-python~=4.7.0.68
numpy~=1.24.2
tensorflow~=2.10.1
torch~=1.13.1+cu117
setuptools~=60.2.0
matplotlib~=3.7.0
torchvision~=0.14.1+cu117
PyYAML~=6.0
requests~=2.28.2
ultralytics~=8.0.39
Pillow~=9.4.0
keras~=2.10.0